refactor: Rename runtime history package to store
