+++ internal/runtime/routes/handler/automation.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/pardnchiu/agenvoy/internal/filesystem/skill"
 	"github.com/pardnchiu/agenvoy/internal/runtime"
-	historyStore "github.com/pardnchiu/agenvoy/internal/runtime/history"
+	historyStore "github.com/pardnchiu/agenvoy/internal/runtime/store"
 	sessionManager "github.com/pardnchiu/agenvoy/internal/session"
 	schedulerTool "github.com/pardnchiu/agenvoy/internal/tools/scheduler"
 )
