+++ internal/runtime/routes/handler/sessionArtifact.go
@@ -11,7 +11,7 @@ import (
 	go_pkg_filesystem_reader "github.com/pardnchiu/go-pkg/filesystem/reader"
 
 	"github.com/pardnchiu/agenvoy/internal/filesystem"
-	historyStore "github.com/pardnchiu/agenvoy/internal/runtime/history"
+	historyStore "github.com/pardnchiu/agenvoy/internal/runtime/store"
 	usagelog "github.com/pardnchiu/agenvoy/internal/session/usage"
 )
