+++ internal/tools/history/action/record.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"time"
 
-	historyStore "github.com/pardnchiu/agenvoy/internal/runtime/history"
+	historyStore "github.com/pardnchiu/agenvoy/internal/runtime/store"
 	toolTypes "github.com/pardnchiu/agenvoy/internal/tools/types"
 )
