+++ internal/tools/history/action/search.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"time"
 
-	historyStore "github.com/pardnchiu/agenvoy/internal/runtime/history"
+	historyStore "github.com/pardnchiu/agenvoy/internal/runtime/store"
 	"github.com/pardnchiu/agenvoy/internal/runtime/torii"
 	sessionHistory "github.com/pardnchiu/agenvoy/internal/session/history"
 	toolTypes "github.com/pardnchiu/agenvoy/internal/tools/types"
